requirements.txt

PyYAML~=6.0.1
streamlit~=1.25.0
click~=8.1.6

--extra-index-url https://download.pytorch.org/whl/cu113
torch
transformers~=4.31.0
streamlit_authenticator
langchain
chromadb==0.3.22
InstructorEmbedding
sentence-transformers